Fully done chapter 1: varargs can be traversed multiple times!
Since the varargs are passed to the function, the function
knows the offset of the first arg. The arg is never consumed,
it still exists as the bytes passed to an fn. This allows
the fn to traverse the args again by calling va_start,
since va_list is just a pointer that points to the next arg.