Write the class struct that every object needs
This class struct points to an objects ctor, dtor
clone and other functions. It allows for dynamic
linkage when calling a function. The object chooses
its own installed function to run on it. Since new
and delete need to know these functions, the class
struct is placed in the new.c file.