Add in the delete fn with reservations in comment
It's not clear why delete frees the ptr returned
by the dtor. Dtor may return a null to preven the
object from being deleted, and now there is a
problem. The other issue is that of the .gitignore.
It needs to be centralized, instead of one for each
chapter, but the regex doesn't seem to work right.