Implement the atom ctor function
This object, an atom is weird because it manages
memory behind the scenes, tricking new and delete
into thinking that the memory they allocated was
not just freed instantly. In general, new and
delete should be the only ones freeing space for
object structs. This is lending itself to ctors
and dtors that cheat the new and delete functions.
When delete thinks it's deleting allocated memory,
it's not doing anything.