comments understanding the node structures
Main lessons learned through inspecting
code: default does not have to be the last
label in a switch case. It can be in the
middle and will still perform as though it
is the last case. Type casting has second
precedence, whereas struct member access
through a pointer has first order precedence.