Introduce the Sine function as a unary op
Very similar to the Inverse operation,
it was super easy to add in the unary
op sin. of course, none of the parsing
is changed, there are no new levels
of binding, the unary operators all
have the highest level. Another restriction
is that none of the op's can be specified
by more than 1 char, since then we'd
have to match against a list of tokens
instead of using a switch statement.